backend/geodata_utils/saih/requirements.txt
pandas==1.4.2
psycopg2-binary==2.9.1
python-decouple==3.6.0